+++ go.mod
@@ -13,7 +13,7 @@ require (
 	github.com/magiconair/properties v1.8.4 // indirect
 	github.com/mitchellh/mapstructure v1.4.1 // indirect
 	github.com/pelletier/go-toml v1.8.1 // indirect
-	github.com/segmentio/ksuid v1.0.3 // indirect
+	github.com/segmentio/ksuid v1.0.3
 	github.com/sethvargo/go-envconfig v0.3.4
 	github.com/spf13/afero v1.5.1 // indirect
 	github.com/spf13/jwalterweatherman v1.1.0 // indirect
